data relation

+++ app/src/main/java/webServer/schools/SchoolRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * SchoolRepository
  */
-public interface SchoolRepository extends JpaRepository<School, Long>{
+public interface SchoolRepository extends JpaRepository<School, Integer>{
 
 	
 }
